Use activity service to get climbing info
We get an activity summary and must use the activity service to get
detailed activity information. Sadly we can't (seem to) batch these
requests so it's quite slow. We then use the detailed activity info to
count up climbs of each category (1-4 and HC) as well as noting the
overall biggest climb and the ones with highest average and maximum
grades. It would be interesting to filter out outliers, such as the
current apparent toughest climb which claims to have a 33% average
grade, but as it's only 130 m I'm skeptical.